Move addonManager files from disco to core
