+++ src/core/addonManager.js
